feat: New option, syncWithHistory, that will manage the history stack
Will default to true.

When false, and before, we would sometimes have situations that would be a little weird for end users. An example of this is when having a slow connection and clicking on first one link and then another before loading has completed. This would before result in two entires in the history stack, meaning that when we come to the page for our second link and then go back we would not go to the page we started at but rather the first link.

Also added in this feat is the possibility to reset the URL in the browser when aborting, something that will be enabled by default when syncWithHistory is true and no custom onAborted is defined.